feat: Update keyboard scroll modifier to support different list types
