+++ patterns/src/main/java/uk/gov/android/ui/patterns/leftalignedscreen/LeftAlignedScreenContentV2.kt
@@ -1,9 +1,6 @@
 package uk.gov.android.ui.patterns.leftalignedscreen
 
 import androidx.compose.foundation.Image
-import androidx.compose.foundation.focusable
-import androidx.compose.foundation.gestures.animateScrollBy
-import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.fillMaxWidth
@@ -15,20 +12,9 @@ import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.MaterialTheme.colorScheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.rememberCoroutineScope
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.focus.FocusRequester
-import androidx.compose.ui.focus.focusRequester
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.vector.ImageVector
-import androidx.compose.ui.input.key.Key
-import androidx.compose.ui.input.key.KeyEventType
-import androidx.compose.ui.input.key.key
-import androidx.compose.ui.input.key.onKeyEvent
-import androidx.compose.ui.input.key.type
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
@@ -37,7 +23,6 @@ import androidx.compose.ui.text.AnnotatedString
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.Dp
 import kotlinx.collections.immutable.ImmutableList
-import kotlinx.coroutines.launch
 import uk.gov.android.ui.componentsv2.R
 import uk.gov.android.ui.componentsv2.button.ButtonTypeV2
 import uk.gov.android.ui.componentsv2.button.GdsButton
@@ -56,6 +41,7 @@ import uk.gov.android.ui.componentsv2.row.RowList
 import uk.gov.android.ui.componentsv2.supportingtext.GdsSupportingText
 import uk.gov.android.ui.componentsv2.warning.GdsWarningText
 import uk.gov.android.ui.theme.dividerThickness
+import uk.gov.android.ui.patterns.utils.scroll.bringIntoView as bringIntoViewV2
 
 internal data class LeftAlignedScreenContentV2(
     val title: String,
@@ -359,39 +345,6 @@ private fun LazyListScope.toAnnotatedText(
  * @param scrollState [LazyListState] represents the list state
  * @return augmented [Modifier]
  */
+@Deprecated("Use uk.gov.android.ui.patterns.utils.scroll.bringIntoView")
 @Composable
-fun Modifier.bringIntoView(scrollState: LazyListState): Modifier {
-    val coroutineScope = rememberCoroutineScope()
-    val interactionSource = remember { MutableInteractionSource() }
-    val focusRequester = remember { FocusRequester() }
-    return this
-        .onKeyEvent {
-            when {
-                it.type == KeyEventType.KeyDown && it.key == Key.DirectionDown &&
-                    scrollState.canScrollForward -> {
-                    coroutineScope.launch {
-                        scrollState.animateScrollBy(
-                            SCROLL_MULTIPLIER * scrollState.layoutInfo.viewportSize.height,
-                        )
-                    }
-                    true
-                }
-
-                it.type == KeyEventType.KeyDown && it.key == Key.DirectionUp &&
-                    scrollState.canScrollBackward -> {
-                    coroutineScope.launch {
-                        scrollState.animateScrollBy(
-                            -SCROLL_MULTIPLIER * scrollState.layoutInfo.viewportSize.height,
-                        )
-                    }
-                    true
-                }
-
-                else -> false
-            }
-        }
-        .focusRequester(focusRequester)
-        .focusable(interactionSource = interactionSource)
-}
-
-private const val SCROLL_MULTIPLIER = 0.4f
+fun Modifier.bringIntoView(scrollState: LazyListState): Modifier = this.bringIntoViewV2(scrollState)
